apps/auth-service/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
python-jose[cryptography]>=3.3
redis>=5.0
